+++ lib/Bitbucket/API/Http/HttpPluginClientBuilder.php
@@ -6,26 +6,56 @@ use Bitbucket\API\Http\Plugin\ApiVersionPlugin;
 use Http\Client\Common\HttpMethodsClient;
 use Http\Client\Common\Plugin;
 use Http\Client\Common\PluginClient;
-use Http\Client\HttpClient;
-use Http\Discovery\HttpClientDiscovery;
 use Http\Discovery\MessageFactoryDiscovery;
+use Http\Discovery\Psr17FactoryDiscovery;
+use Http\Discovery\Psr18ClientDiscovery;
 use Http\Message\MessageFactory;
+use Psr\Http\Client\ClientInterface;
+use Psr\Http\Message\RequestFactoryInterface;
+use Psr\Http\Message\StreamFactoryInterface;
 
 class HttpPluginClientBuilder
 {
-    /** @var HttpClient */
+    /** @var ClientInterface */
     private $httpClient;
     /** @var HttpMethodsClient|null */
     private $pluginClient;
-    /** @var MessageFactory */
-    private $messageFactory;
+    /** @var MessageFactory|RequestFactoryInterface */
+    private $requestFactory;
+    /** @var StreamFactoryInterface */
+    private $streamFactory;
     /** @var Plugin[] */
     private $plugins = [];
 
-    public function __construct(HttpClient $httpClient = null, MessageFactory $messageFactory = null)
+    /**
+     * @param MessageFactory|RequestFactoryInterface|null $requestFactory
+     */
+    public function __construct(ClientInterface $httpClient = null, $requestFactory = null, StreamFactoryInterface $streamFactory = null)
     {
-        $this->httpClient = $httpClient ?: HttpClientDiscovery::find();
-        $this->messageFactory = $messageFactory ?: MessageFactoryDiscovery::find();
+        $requestFactory = $requestFactory ?? Psr17FactoryDiscovery::findRequestFactory();
+        if ($requestFactory instanceof MessageFactory) {
+            // Use same format as symfony/deprecation-contracts.
+            @trigger_error(sprintf(
+                'Since %s %s: %s is deprecated, use %s instead.',
+                'private-packagist/bitbucket-api',
+                '2.2.0',
+                '\Http\Message\MessageFactory',
+                RequestFactoryInterface::class
+            ), \E_USER_DEPRECATED);
+        } elseif (!$requestFactory instanceof RequestFactoryInterface) {
+            /** @var mixed $requestFactory value unknown; set to mixed, prevent PHPStan complaining about guard clauses */
+            throw new \TypeError(sprintf(
+                '%s::__construct(): Argument #2 ($requestFactory) must be of type %s|%s, %s given',
+                self::class,
+                '\Http\Message\MessageFactory',
+                RequestFactoryInterface::class,
+                is_object($requestFactory) ? get_class($requestFactory) : gettype($requestFactory)
+            ));
+        }
+
+        $this->httpClient = $httpClient ?: Psr18ClientDiscovery::find();
+        $this->requestFactory = $requestFactory;
+        $this->streamFactory = $streamFactory ?? Psr17FactoryDiscovery::findStreamFactory();
     }
 
     /**
@@ -79,7 +109,8 @@ class HttpPluginClientBuilder
         if (!$this->pluginClient) {
             $this->pluginClient = new HttpMethodsClient(
                 new PluginClient($this->httpClient, $this->plugins),
-                $this->messageFactory
+                $this->requestFactory,
+                $this->streamFactory
             );
         }
 
@@ -88,9 +119,20 @@ class HttpPluginClientBuilder
 
     /**
      * @return MessageFactory
+     * @deprecated Use getRequestFactory instead. message will be removed with 3.0
      */
     public function getMessageFactory()
     {
-        return $this->messageFactory;
+        return $this->requestFactory instanceof MessageFactory
+            ? $this->requestFactory
+            : MessageFactoryDiscovery::find();
+    }
+
+    /**
+     * @return RequestFactoryInterface
+     */
+    public function getRequestFactory()
+    {
+        return $this->requestFactory;
     }
 }