+++ lib/Bitbucket/API/Http/Plugin/ApiVersionPlugin.php
@@ -5,6 +5,7 @@ namespace Bitbucket\API\Http\Plugin;
 use Http\Client\Common\Plugin;
 use Http\Promise\Promise;
 use Psr\Http\Message\RequestInterface;
+use Psr\Http\Message\ResponseInterface;
 
 class ApiVersionPlugin implements Plugin
 {
@@ -26,7 +27,7 @@ class ApiVersionPlugin implements Plugin
      * @param callable $next Next middleware in the chain, the request is passed as the first argument
      * @param callable $first First middleware in the chain, used to to restart a request
      *
-     * @return Promise Resolves a PSR-7 Response or fails with an Http\Client\Exception (The same as HttpAsyncClient).
+     * @return Promise<ResponseInterface> Resolves a PSR-7 Response or fails with an Http\Client\Exception (The same as HttpAsyncClient).
      */
     protected function doHandleRequest(RequestInterface $request, callable $next, callable $first)
     {